airpollpredictor/requirements.txt
lightgbm~=3.3.5
optuna==3.1.1
pmdarima~=2.0.3
dvc
ipywidgets~=8.0.6
ipython~=8.12.0
joblib
matplotlib~=3.7.1
numpy~=1.24.2
pandas~=2.0.0
requests~=2.28.2
pathspec
fastapi~=0.95.0
pydantic~=1.10.7
uvicorn
PyYAML~=6.0
statsmodels~=0.13.5
scipy~=1.10.0
scikit-learn~=1.2.1
seaborn~=0.12.2
progress
awscli
dvc_s3